3-10 views.py 파일 분리 완료
