+++ djangoProject/urls.py
@@ -1,10 +1,10 @@
 from django.contrib import admin
 from django.urls import path, include
-from pybo import views
+from pybo.views import base_views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('pybo/', include('pybo.urls')),
-    path('', views.index, name="index"),
+    path('', base_views.index, name="index"),
     path('common/', include('common.urls')),
 ]
